requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==22.1
django-modeltranslation==0.18.9
django-rq==2.4.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.5
easy-thumbnails==2.8.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
polib==1.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
rq==1.13.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
